Fix trivial Windows support issues (#2005)
* Initial int_ fixes

* Fix permission errors

* linesep fixes

* Fix test_tabular_msa

* Fix int32 out of range errors

* Improve lladser from casting error to buffer dtype mismatch

* Fix degap tests for grammared sequence

* Fix dtype issues for sequence tests

* Fix all errors in test_sequence

* Fix errors in metadata test_missing

* Fixed line ending issues in io

* Change formatting

* Fix more line separator issues in io/util

* Fix open encoding issue

+++ skbio/stats/_subsample.py
@@ -117,7 +117,7 @@ def isubsample(items, maximum, minimum=1, buf_size=1000, bin_f=None):
             return True
 
     # buffer some random values
-    random_values = np.random.randint(0, sys.maxsize, buf_size)
+    random_values = np.random.randint(0, sys.maxsize, buf_size, dtype=np.int64)
     random_idx = 0
 
     result = defaultdict(list)
@@ -130,7 +130,7 @@ def isubsample(items, maximum, minimum=1, buf_size=1000, bin_f=None):
         random_value = random_values[random_idx]
         random_idx += 1
         if random_idx >= buf_size:
-            random_values = np.random.randint(0, sys.maxsize, buf_size)
+            random_values = np.random.randint(0, sys.maxsize, buf_size, dtype=np.int64)
             random_idx = 0
 
         # push our item on to the heap and drop the smallest if necessary
@@ -225,7 +225,7 @@ def subsample_counts(counts, n, replace=False):
         raise ValueError("n cannot be negative.")
 
     counts = np.asarray(counts)
-    counts = counts.astype(int, casting="safe")
+    counts = counts.astype(np.int64, casting="safe")
 
     if counts.ndim != 1:
         raise ValueError("Only 1-D vectors are supported.")
